+++ R/word2vec.R
@@ -1,7 +1,7 @@
 #' Word2vec model
 #' 
 #' Train a Word2vec model (Mikolov et al., 2023) in different architectures on a [quanteda::tokens] object.
-#' @param x a [quanteda::tokens] object.
+#' @param x a [quanteda::tokens] or [quanteda::tokens_xptr] object.
 #' @param dim the size of the word vectors.
 #' @param type the architecture of the model; either "cbow" (continuous back of words) or "skip-gram".
 #' @param min_count the minimum frequency of the words. Words less frequent than 
@@ -15,7 +15,8 @@
 #' @param ns_size the size of negative samples. Only used when `use_ns = TRUE`.
 #' @param sample the rate of sampling of words based on their frequency. Sampling is 
 #'   disabled when `sample = 1.0`
-#' @param normalize if `TRUE`, normalize the vectors in `values` and `weights`.
+#' @param tolower lower-case all the tokens before fitting the model.
+#' @param model a trained Word2vec model; if provided, its word vectors are updated for `x`.
 #' @param verbose if `TRUE`, print the progress of training.
 #' @param ... additional arguments.
 #' @returns Returns a textmodel_wordvector object with the following elements:
@@ -29,6 +30,7 @@
 #'   \item{alpha}{the initial learning rate.}
 #'   \item{use_ns}{the use of negative sampling.}
 #'   \item{ns_size}{the size of negative samples.}
+#'   \item{min_count}{the value of min_count.}
 #'   \item{concatenator}{the concatenator in `x`.}
 #'   \item{call}{the command used to execute the function.}
 #'   \item{version}{the version of the wordvector package.}
@@ -64,8 +66,8 @@
 #' }
 textmodel_word2vec <- function(x, dim = 50, type = c("cbow", "skip-gram"), 
                                min_count = 5L, window = ifelse(type == "cbow", 5L, 10L), 
-                               iter = 10L, alpha = 0.05, use_ns = TRUE, ns_size = 5L, 
-                               sample = 0.001, normalize = TRUE,
+                               iter = 10L, alpha = 0.05, model = NULL, 
+                               use_ns = TRUE, ns_size = 5L, sample = 0.001, tolower = TRUE,
                                verbose = FALSE, ...) {
     UseMethod("textmodel_word2vec")
 }
@@ -76,9 +78,10 @@ textmodel_word2vec <- function(x, dim = 50, type = c("cbow", "skip-gram"),
 #' @method textmodel_word2vec tokens
 textmodel_word2vec.tokens <- function(x, dim = 50L, type = c("cbow", "skip-gram"), 
                                       min_count = 5L, window = ifelse(type == "cbow", 5L, 10L), 
-                                      iter = 10L, alpha = 0.05, use_ns = TRUE, ns_size = 5L, 
-                                      sample = 0.001, normalize = TRUE,
-                                      verbose = FALSE, ..., old = FALSE) {
+                                      iter = 10L, alpha = 0.05, model = NULL, 
+                                      use_ns = TRUE, ns_size = 5L, sample = 0.001, tolower = TRUE,
+                                      verbose = FALSE, ..., 
+                                      normalize = FALSE, old = FALSE) {
     
     type <- match.arg(type)
     dim <- check_integer(dim, min = 2)
@@ -90,23 +93,37 @@ textmodel_word2vec.tokens <- function(x, dim = 50L, type = c("cbow", "skip-gram"
     alpha <- check_double(alpha, min = 0)
     sample <- check_double(sample, min = 0)
     normalize <- check_logical(normalize)
+    tolower <- check_logical(tolower)
     verbose <- check_logical(verbose)
-
+    
+    if (normalize)
+        .Deprecated(msg = "normalize is deprecated. Use as.matrix(x, normalize = TRUE) instead.")
+    
     type <- match(type, c("cbow", "skip-gram"))
     if (old)
         type <- type * 10
     
-    # NOTE: use tokens_xptr?
+    if (!is.null(model)) {
+        if (!identical(class(model), "textmodel_wordvector"))
+            stop("model must be a trained textmodel_wordvector")
+        if (!identical(model$dim, dim))
+            stop("model must be trained with dim = ", dim)
+    }
+    
+    x <- as.tokens_xptr(x)
+    if (tolower)
+        x <- tokens_tolower(x)
     x <- tokens_trim(x, min_termfreq = min_count, termfreq_type = "count")
-    result <- cpp_w2v(as.tokens(x), attr(x, "types"), 
-                      minWordFreq = min_count,
-                      size = dim, window = window,
+    
+    result <- cpp_w2v(x, size = dim, window = window,
                       sample = sample, withHS = !use_ns, negative = ns_size, 
                       threads = get_threads(), iterations = iter,
-                      alpha = alpha, type = type, normalize = normalize, verbose = verbose)
+                      alpha = alpha, type = type, normalize = normalize, model = model,
+                      verbose = verbose)
     if (!is.null(result$message))
         stop("Failed to train word2vec (", result$message, ")")
-
+    
+    result$min_count <- min_count
     result$concatenator <- meta(x, field = "concatenator", type = "object")
     result$call <- try(match.call(sys.function(-1), call = sys.call(-1)), silent = TRUE)
     result$version <- utils::packageVersion("wordvector")
@@ -120,7 +137,7 @@ word2vec <- function(...) {
 
 #' Print method for trained word vectors
 #' @param x for print method, the object to be printed
-#' @param ... unused
+#' @param ... not used.
 #' @method print textmodel_wordvector
 #' @keywords internal
 #' @return an invisible copy of `x`. 
@@ -155,9 +172,15 @@ print.textmodel_docvector <- function(x, ...) {
 #'
 #' Extract word vectors from a `textmodel_wordvector` or `textmodel_docvector` object.
 #' @param x a `textmodel_wordvector` or `textmodel_docvector` object.
-#' @param ... not used
-#' @return a matrix that contain the word vectors in rows
+#' @param normalize if `TRUE`, returns normalized word vectors.
+#' @param ... not used.
+#' @return a matrix that contain the word vectors in rows.
 #' @export
-as.matrix.textmodel_wordvector <- function(x, ...){
+as.matrix.textmodel_wordvector <- function(x, normalize = TRUE, ...){
+    normalize <- check_logical(normalize)
+    if (normalize) {
+        v <- sqrt(rowSums(x$values ^ 2) / ncol(x$values))
+        x$values <- x$values / v
+    }
     return(x$values) 
 }